+++ tests/index.test.ts
@@ -11,13 +11,16 @@ async function test() {
 
   const x = await conn
     .query<{ username: string; id: string }>()
-    .select("*")
+    .select(["id"], true)
     .from("users")
-    .where("username", "qsd")
-    .order("username", "ASC")
+    .where("id", "a7f17514-e434-4142-bc68-9360ce26df6e")
+    .or("id", "f7109fd2-6678-4e5f-b9ec-c3c84f574f72")
     .exec();
 
+  // const d = await conn.query().drop("books", "database").exec();
+
   console.log(x);
+
   console.log(conn.threadId);
 
   conn.end();